✨ feat: create schemas routes
