+++ src/app/routes/index.ts
@@ -3,6 +3,7 @@ import { Router } from "express";
 import loginRouter from "./login";
 import usersRouter from "./users";
 import { projectsRouter } from "./projects";
+import { schemasRouter } from "./schemas";
 
 const routes = (): Router => {
   const router = Router();
@@ -11,6 +12,7 @@ const routes = (): Router => {
   usersRouter(router);
 
   router.use("/projects", projectsRouter);
+  router.use("/projects", schemasRouter);
 
   return router;
 };
